+++ cudajit.mli
@@ -790,3 +790,43 @@ module Event : sig
       {{:https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__STREAM.html#group__CUDA__STREAM_1g6a898b652dfc6aa1d5c8d97062618b2f}
         cuStreamWaitEvent}. *)
 end
+
+(** This module builds on top of functionality more directly exposed by {!Event}. It optimizes
+    resource management for use-cases where events are not reused: there's only one call to
+    [record], and it's immediately after [create]. *)
+module Delimited_event : sig
+  type t
+  (** An delimited event encapsulates {!Event.t} and is owned by a stream. It records its owner at
+      creation, and gets destroyed when either it or its owner are synchronized (or if neither
+      happens, when it is garbage-collected). *)
+
+  val record :
+    ?blocking_sync:bool ->
+    ?enable_timing:bool ->
+    ?interprocess:bool ->
+    ?external_:bool ->
+    Stream.t ->
+    t
+  (** Combines {!Event.create} and {!Event.record} to create an event owned by the given stream. *)
+
+  val is_destroyed : t -> bool
+  (** Returns true if the delimited event is already released via a call to
+      {{:https://docs.nvidia.com/cuda/cuda-driver-api/group__CUDA__EVENT.html#group__CUDA__EVENT_1g593ec73a8ec5a5fc031311d3e4dca1ef}
+        cuEventDestroy}. The event can be released via either {!synchronize}, or
+      {!Stream.synchronize}. *)
+
+  val elapsed_time : start:t -> end_:t -> float
+  (** See {!Event.elapsed_time}. [elapsed_time ~start ~end_] raises [Invalid_argument] if either
+      [start] or [end_] is already destroyed. *)
+
+  val query : t -> bool
+  (** See {!Event.query}. [query event] raises [Invalid_argument] if [event] is already
+      destroyed. *)
+
+  val synchronize : t -> unit
+  (** See {!Event.synchronize}. [synchronize event] is a no-op if [is_destroyed event] is true. *)
+
+  val wait : ?external_:bool -> Stream.t -> t -> unit
+  (** See {!Event.wait}. [wait stream event] raises [Invalid_argument] if [event] is already
+      destroyed. *)
+end